requirements.txt
absl-py==2.3.1
accelerate==0.29.3
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.7.2
anyio==4.11.0
attrs==25.3.0
av==15.1.0
bitsandbytes==0.45.5
black==25.9.0
certifi==2025.8.3
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
DataProperty==1.1.0
datasets==4.1.1
decord==0.6.0
dill==0.4.0
distlib==0.4.0
distro==1.9.0
einops==0.6.1
einops-exts==0.0.4
et_xmlfile==2.0.0
evaluate==0.4.6
fastapi==0.118.0
ffmpy==0.6.1
filelock==3.13.1
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2024.6.1
ftfy==6.3.1
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.45
gradio==4.16.0
gradio_client==0.8.1
grpcio==1.75.1
h11==0.14.0
hf-xet==1.1.10
hf_transfer==0.1.9
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.35.3
identify==2.6.14
idna==3.10
importlib_resources==6.5.2
iopath==0.1.10
isort==6.1.0
Jinja2==3.1.4
jiter==0.11.0
joblib==1.5.2
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
latex2mathml==3.78.1
latex2sympy2==1.9.1
lightning-bolts==0.7.0
lightning-utilities==0.15.2
-e git+https://github.com/haotian-liu/LLaVA.git@c121f0432da27facab705978f83c4ada465e46fd#egg=llava
lmdb==1.7.3
-e git+https://github.com/EvolvingLMMs-Lab/lmms-eval@bbd344ee52179e48efd4a0e28a6d14490d7e4dbe#egg=lmms_eval
loguru==0.7.3
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==2.1.5
matplotlib==3.10.6
mbstrdecoder==1.1.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.6.0
networkx==3.3
nltk==3.9.2
nodeenv==1.9.1
numexpr==2.13.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
openai==2.0.1
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pathvalidate==3.3.1
peft==0.11.1
pillow==10.4.0
platformdirs==4.4.0
portalocker==3.2.0
pre_commit==4.3.0
propcache==0.3.2
protobuf==4.25.8
psutil==7.1.0
pyarrow==21.0.0
pybind11==3.0.1
pycocoevalcap==1.2
pycocotools==2.0.10
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.1.10
pytorch-lightning==1.9.5
pytz==2025.2
PyYAML==6.0.3
qwen-vl-utils==0.0.14
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rfc3986==1.5.0
rich==14.1.0
rpds-py==0.27.1
ruff==0.13.2
sacrebleu==2.5.1
safetensors==0.6.2
scikit-learn==1.2.2
scipy==1.16.2
semantic-version==2.10.0
sentence-transformers==5.1.1
sentencepiece==0.1.99
sentry-sdk==2.39.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlitedict==2.1.0
starlette==0.48.0
svgwrite==1.4.3
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tenacity==8.3.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.11.0
timm==1.0.9
tokenizers==0.15.2
tomlkit==0.12.0
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchmetrics==1.8.2
torchvision==0.16.2+cu121
tqdm==4.67.1
tqdm-multiprocess==0.0.11
transformers==4.37.2
transformers-stream-generator==0.0.5
triton==2.1.0
typepy==1.3.4
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
virtualenv==20.34.0
wandb==0.22.1
wavedrom==2.0.3.post3
wcwidth==0.2.14
websockets==11.0.3
Werkzeug==3.1.3
xxhash==3.6.0
yacs==0.1.8
yarl==1.20.1
yt-dlp==2025.9.26
zss==1.2.0
zstandard==0.25.0
